+++ crates/gateway-types/src/reply.rs
@@ -773,26 +773,3 @@ mod tests {
         }
     }
 }
-
-// Types used for web socket subscription events
-pub mod ws_subscriptions {
-    use super::Block;
-    use serde::Serialize;
-
-    #[derive(Debug, Clone, Serialize)]
-    pub struct SubscriptionSyncEvent {
-        pub block: Box<Block>,
-    }
-
-    #[derive(Debug, Clone, Serialize)]
-    pub struct SubscriptionNewHeadEvent {
-        pub block: Box<Block>,
-    }
-
-    /// Events and queries emitted by L2 sync process.
-    #[derive(Debug, Clone, Serialize)]
-    pub enum SubscriptionEvent {
-        Sync(SubscriptionSyncEvent),
-        NewHead(SubscriptionNewHeadEvent),
-    }
-}
